refactor: rename package - provider => manager
Related to: #87
